+++ src/jk_logging/_wrapMain.py
@@ -7,6 +7,7 @@ import sys
 from .ExceptionInChildContextException import ExceptionInChildContextException
 from .AbstractLogger import AbstractLogger
 from .ConsoleLogger import ConsoleLogger
+from .fmt.LogMessageFormatter import LogMessageFormatter, DEFAULT_LOG_MESSAGE_FORMATTER
 from .fmt.ColoredLogMessageFormatter import ColoredLogMessageFormatter, COLOR_LOG_MESSAGE_FORMATTER
 
 
@@ -15,7 +16,7 @@ from .fmt.ColoredLogMessageFormatter import ColoredLogMessageFormatter, COLOR_LO
 
 
 
-def wrapMain(log:AbstractLogger = None):
+def wrapMain(log:AbstractLogger = None, bColor:bool = True):
 
 	class _LogCtx(object):
 
@@ -41,7 +42,10 @@ def wrapMain(log:AbstractLogger = None):
 	#
 
 	if log is None:
-		log = ConsoleLogger.create(logMsgFormatter=COLOR_LOG_MESSAGE_FORMATTER)
+            if bColor:
+                log = ConsoleLogger.create(logMsgFormatter=COLOR_LOG_MESSAGE_FORMATTER)
+            else:
+                log = ConsoleLogger.create(logMsgFormatter=DEFAULT_LOG_MESSAGE_FORMATTER)
 	else:
 		assert isinstance(log, AbstractLogger)